requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
channels==4.0.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-redis==5.3.0
django-timezone-field==5.1
factory-boy==3.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kombu==5.3.1
mysqlclient==2.2.0
prompt-toolkit==3.0.39
pytest==7.4.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
sentry-sdk==1.28.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
